Moving discord button, cleanup up styles
