+++ src/views/Launch.tsx
@@ -1,19 +1,26 @@
 import React, { useState, CSSProperties } from "react";
-import { Text, Stack, Button, IconButton, Image, ImageCoverStyle, ImageFit, Separator } from "office-ui-fabric-react";
+import { Text, Stack, Button, IconButton, Image, ImageCoverStyle, ImageFit } from "office-ui-fabric-react";
 import { Route } from "react-router";
 import { NavLink } from 'react-router-dom';
 import styled from 'styled-components';
 
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import HoverBox from '../components/HoverBox';
 import { Images } from "../common/const";
 import { Participants } from "../views/launch/participants";
-import { Signup } from "../views/launch/signup";
+import { Signin } from "../views/launch/signin";
 
+const FaIconStyle: CSSProperties = {
+    color: "white",
+    height: "20px",
+    width: "20px",
+    paddingLeft: "10px"
+};
 export const Launch: React.StatelessComponent<any> = ({ match }: any) => {
     return (
         <Stack>
             <Route path={`${match.path}/participants`} component={Participants} />
-            <Route path={`${match.path}/signup`} component={Signup} />
+            <Route path={`${match.path}/sign`} component={Signin} />
             <Route exact path={match.path} component={LaunchViewSelection} />
         </Stack>
     );
@@ -27,7 +34,7 @@ const LaunchViewSelection = () => {
                     <Image width="100%" height="400px" src={Images.launchAppsHero} coverStyle={ImageCoverStyle.landscape} imageFit={ImageFit.cover} />
                     <Stack style={{margin: "20px"}}>
 
-                        <Text style={{fontFamily: "Segoe UI", fontSize: "40px", fontWeight: "lighter"}}>// Launch</Text>
+                        <Text style={{fontFamily: "Segoe UI", fontSize: "30px", fontWeight: "lighter"}}>// Launch</Text>
 
                         <Text  style={{marginTop: "10px", fontWeight: 500}} variant="xLarge">An annual event where a community of Windows App developers release their UWP-related projects</Text>
                         <Text style={{marginTop: "10px"}} variant="mediumPlus">Our Discord server provides direct, 2 way user feedback from users, and a place for newbies to ask questions and learn from those with more experience, creating the perfect environment for apps to grow into something more</Text>
@@ -37,7 +44,10 @@ const LaunchViewSelection = () => {
             <Stack horizontal wrap horizontalAlign="center" tokens={{childrenGap: 25}}>
                 <LaunchCard header="Participating apps" description="See which apps are participating in Launch 2020" path="/launch/participants" />
 
-                <LaunchCard header="Submit your app" description="Want to Launch your app with the community?" buttonText="Sign up now" path="/launch/signup" buttonDisabled={true} />                    
+                <LaunchCard header="Submit your app" description="Want to Launch your app with the community?" path="/launch/signin" buttonStyle={{paddingTop: "25px", paddingBottom: "25px", marginLeft: "10px"}} buttonDisabled={false}>
+                    <Text>Sign in </Text>
+                    <FontAwesomeIcon style={FaIconStyle} icon={["fab", "discord"]} />
+                </LaunchCard>       
             </Stack>
         </Stack>
     );
@@ -49,26 +59,27 @@ interface ILaunchCardProps {
     path: string;
     buttonText?: string;
     buttonDisabled?: boolean;
+    buttonStyle?: CSSProperties;
 };
 const LaunchCardStyle = styled(HoverBox)`
     max-height: 500px;
-    max-width: 300px;
+    max-width: 350px;
     padding: 50px;
     :hover {
                 pointer: cursor;
         };
     `;
 
-const LaunchCard: React.FC<ILaunchCardProps> = (props: ILaunchCardProps) => {
+const LaunchCard: React.FunctionComponent<ILaunchCardProps> = (props: React.PropsWithChildren<ILaunchCardProps>) => {
     return (
         <LaunchCardStyle>
-            <Stack>
-                <Stack horizontal tokens={{ childrenGap: 5 }}>
+            <Stack tokens={{childrenGap: 5}}>
+                <Stack horizontal tokens={{ childrenGap: 5 }} verticalAlign="center">
                     <Text variant="xLarge">{props.header}</Text>
                     <NavLink to={(props.buttonDisabled? window.location.pathname : props.path)}>
                         {
-                            props.buttonText ?
-                                <Button disabled={props.buttonDisabled} style={{ marginLeft: "10px" }} text={props.buttonText} />
+                            props.buttonText || props.children ?
+                                <Button primary disabled={props.buttonDisabled} style={props.buttonStyle} text={props.buttonText}>{props.children}</Button>
                                 :
                                 <IconButton disabled={props.buttonDisabled} primary iconProps={{ iconName: 'Go' }} />
                         }
