+++ src/views/launch/signin.tsx
@@ -1,0 +1,41 @@
+import { Text, Stack, Button, BaseButton } from "office-ui-fabric-react";
+import React, { useState, CSSProperties } from "react";
+import { NavLink } from 'react-router-dom';
+
+export const Signin = () => {
+    const [signedIn, setSignedIn] = useState(false);
+
+    return (
+        <Stack>
+
+        </Stack>
+    )
+};
+
+
+
+function SignIntoDiscord_Clicked() {
+    /**
+     * TODO: Basic sign in 
+     * https://discordapp.com/developers/docs/topics/oauth2
+     * 
+     * 
+     * Generate unique id
+     * Establish WebSocket to backend with id
+     * Open discord auth endpoint in new tab, pass id as state param
+     * Wait for user to authenticate
+     * Set up discord auth redirect URI on backend to ping "ok" back to the page via the websocket, using state to ping the right socket
+     * Attempt to auto close the redirect auth page
+     * Use localstorage to keep the user logged in 
+     */
+}
+
+
+/**
+ * Todo: Post-sign in
+ * 
+ * Redirect to a /me/ page to manage apps
+ * Use discord.js to verify the user is in the server
+ * Ask the user to join if not (provide link)
+ * Display their name and icon on the page
+ */
